requirements.txt
# Core ML Libraries
scikit-learn>=1.3.0
xgboost>=2.0.0
lightgbm>=4.1.0
pandas>=2.1.0
numpy>=1.24.0

# LLM Integration
google-generativeai>=0.3.0

# Kaggle API (for real mode)
kaggle>=1.5.16

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
tqdm>=4.66.0

# Logging & Monitoring
structlog>=23.2.0

# Parallel Processing
joblib>=1.3.0
